Created Auth Provider
