+++ src/index.js
@@ -5,11 +5,18 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { ReactSession } from 'react-client-session';
+import { AuthContextProvider } from './Auth/authContext';
 
 ReactSession.setStoreType("sessionStorage");
 ReactSession.set("username", "User");
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<BrowserRouter><App /></BrowserRouter>);
+root.render(
+    <AuthContextProvider>
+        <BrowserRouter>
+            <App />
+        </BrowserRouter>
+    </AuthContextProvider>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
