+++ taskLIst.txt
@@ -3,10 +3,14 @@ To Do List
 - Create a 404 Not Found Component [PRDEV-008]
 - Find a better way to store all CSS files [PRDEV-010]
 - Create a common way to handle all errors while making fetch requests [PRDEV-011]
+
+-- Epic related to Auth
 - Create a logout button if the user is logged in. [PRDEV-025]
 - Also ensure that the user should be logged-in in every page. [PRDEV-026]
 - Create update functionality for the User page. [PRDEV-027]
-- Bug: User state is getting lost on page refresh.
+- Bug: User state is getting lost on page refresh. [PRDEV-028]
+- 'Hi User' should be replaced by a login button [PRDEV-029]
+- Create an AuthProvider and keep it central [PRDEV-030]
